more of import books backend, sort of doing TDD which is quite nice actually

+++ src/Domain.LinnApps/ImportBooks/ImpBookInvoiceDetail.cs
@@ -8,6 +8,6 @@
 
         public string InvoiceNumber { get; set; }
 
-        public int InvoiceValue { get; set; }
+        public decimal InvoiceValue { get; set; }
     }
 }
